requirements.txt
#ExamenBloc2/requirements.txt
sqlalchemy # Pour l'ORM
sqlalchemy-utils # Pour les types de données supplémentaires dans SQLAlchemy
psycopg[binary] # Pour PostgreSQL à voir si nécessaire car la base de données est MySQL
pymysql # Pour MySQL avec SQLAlchemy
cryptography # Pour la sécurité et le chiffrement

jupyter # Pour les notebooks interactifs
ipykernel # Pour exécuter des notebooks Jupyter

pandas # Pour la manipulation des données
matplotlib # Pour la visualisation des données
pyarrow # Pour la gestion des fichiers Parquet
fastparquet # Pour la gestion des fichiers Parquet

kafka-python>=2.0.2 # Pour interagir avec Apache Kafka

python-dotenv # Pour charger les variables d'environnement depuis un fichier .env
codecarbon # Pour le suivi de l'empreinte carbone des programmes Python
tabulate # Pour afficher des tableaux joliment formatés dans la console
pathlib # Pour la gestion des chemins de fichiers

scikit-learn # Pour les algorithmes de machine learning
joblib # Pour la sérialisation des modèles